+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: spark <spark@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/28 16:38:30 by spark             #+#    #+#             */
-/*   Updated: 2020/09/28 16:57:09 by spark            ###   ########.fr       */
+/*   Updated: 2020/09/29 14:27:38 by spark            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(void *p, int tg, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
