+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: spark <spark@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/28 15:26:43 by spark             #+#    #+#             */
-/*   Updated: 2020/09/28 16:49:55 by spark            ###   ########.fr       */
+/*   Updated: 2020/09/29 14:27:48 by spark            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memcpy(void *dest, void *src, size_t n)
 {
 	void	*rt;
-	int		i;
+	size_t	i;
 
 	if (!dest && !src)
 		return (0);
